+++ src/XRoadLib/Schema/TypeDefinition.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
-using XRoadLib.Extensions;
 
 namespace XRoadLib.Schema
 {
@@ -28,7 +26,7 @@ namespace XRoadLib.Schema
 
         public TypeDefinition(Type type)
         {
-            Documentation = type.GetTypeInfo().GetXRoadTitles().Where(title => !string.IsNullOrWhiteSpace(title.Item2)).ToArray();
+            Documentation = new DocumentationDefinition(type.GetTypeInfo());
             Type = type;
         }
     }