+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 1.0.0 - unreleased
+* Remove unnecessary XML namespace prefix from WSDL.
+* Add X-Road specific documentation elements.
+
 #### 1.0.0-beta042 - 22.02.2017
 * Fix NuGet publishing.
 * Remove duplicate entries from API reference.